requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-openai
langchain-community
PyMuPDF